requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
prometheus-fastapi-instrumentator
weasyprint
pydantic
requests
pytest
pytest-cov
httpx
pytest-asyncio
pytest-mock
pytest-httpx
pytest-asyncmock
greenlet
sqlalchemy-asyncpg
pydantic-settings